Update Current props

+++ Weather/Models/Current.swift
@@ -22,7 +22,7 @@ struct Current: Codable {
   var condition: String { ConditionType.classifyCondition(by: weather[0].id).rawValue }
   /// Date object
   var date: Date { Date(timeIntervalSince1970: dt) }
-  var timezone: Double?
+  var timezone: Int?
 }
 
 extension Current: Equatable {
